move spatial joins out of merge cities and create dummy MAZ parcels
